refactor branch_age tests
